Revert traverser factory interface
